@@ -1,7 +1,7 @@
 package sperka.pl.bookcase.server.helpers;
 
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 import java.time.Instant;
 import java.time.format.DateTimeParseException;
 
