@@ -1,5 +1,6 @@
 package sperka.pl.bookcase.server.resource;
 
+import jakarta.annotation.security.PermitAll;
 import lombok.extern.slf4j.Slf4j;
 import sperka.pl.bookcase.server.auth.JwtUserPrincipal;
 import sperka.pl.bookcase.server.dto.GenericResponseDto;
@@ -8,14 +9,13 @@ import sperka.pl.bookcase.server.dto.LoginResponseDto;
 import sperka.pl.bookcase.server.service.AuthService;
 import sperka.pl.bookcase.server.service.UserService;
 
-import javax.annotation.security.PermitAll;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.ws.rs.*;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.SecurityContext;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 
 @ApplicationScoped
 @Path( "/api/auth" )
