@@ -1,5 +1,6 @@
 package sperka.pl.bookcase.server.resource;
 
+import jakarta.annotation.security.RolesAllowed;
 import sperka.pl.bookcase.server.auth.Roles;
 import sperka.pl.bookcase.server.dto.BookDto;
 import sperka.pl.bookcase.server.dto.BookFilterDto;
@@ -8,12 +9,11 @@ import sperka.pl.bookcase.server.exceptions.ValidationException;
 import sperka.pl.bookcase.server.service.BookDatabaseSynchronizationService;
 import sperka.pl.bookcase.server.service.BookService;
 
-import javax.annotation.security.RolesAllowed;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import java.util.List;
 
 @ApplicationScoped
