@@ -2,8 +2,8 @@ package sperka.pl.bookcase.server.validators.annotations;
 
 import sperka.pl.bookcase.server.validators.BookSignatureNotExistsValidator;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
